Move log method to LoggingAdapter
