+++ wookiee-libs/src/main/scala/com/oracle/infy/wookiee/logging/LoggingAdapter.scala
@@ -3,6 +3,7 @@ package com.oracle.infy.wookiee.logging
 import com.oracle.infy.wookiee.actor.WookieeScheduler
 import org.slf4j.{Logger, LoggerFactory}
 
+import java.util.ServiceLoader
 import java.util.logging.Level
 import java.util.logging.Level._
 import scala.concurrent.{ExecutionContext, Future}
@@ -46,5 +47,43 @@ trait LoggingAdapter {
       }
     }
     tried
+  }
+}
+
+object LoggingAdapter {
+
+  // Call this to see where the logging classes are coming from
+  def printLoggingClasses(): Unit = {
+    println("Logging classpath and class information for diagnostics...")
+    logClassInformation("org.slf4j.LoggerFactory")
+    logClassInformation("ch.qos.logback.classic.LoggerContext")
+    logClassInformation("ch.qos.logback.core.util.StatusPrinter")
+    logSLF4JServiceProvider()
+  }
+
+  def logClassInformation(className: String): Unit = {
+    try {
+      val clazz = Class.forName(className)
+      val classLoader = clazz.getClassLoader
+      val location = classLoader.getResource(className.replace('.', '/') + ".class")
+      println(s"Class: $className, ClassLoader: $classLoader, Location: $location")
+    } catch {
+      case e: ClassNotFoundException => println(s"Class $className not found: ${e.getMessage}")
+      case e: Exception              => println(s"Error loading class $className: ${e.getMessage}")
+    }
+  }
+
+  def logSLF4JServiceProvider(): Unit = {
+    println("SLF4J Service Providers:")
+    try {
+      val serviceLoader = ServiceLoader.load(classOf[org.slf4j.spi.SLF4JServiceProvider])
+      serviceLoader.forEach { provider =>
+        println(
+          s"Provider: ${provider.getClass.getName}, sl4fj max version supported: ${provider.getRequestedApiVersion}"
+        )
+      }
+    } catch {
+      case e: Exception => println(s"Error loading SLF4J Service Providers: ${e.getMessage}")
+    }
   }
 }
